Refactoring - merged Recipebook read and write service implementation
there is no need for two classes
